# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/{universe}/universe-creation/equipment-types/jewellery-definitions": {
      "get": {
        "operationId": "getDefaultJewelleryDefinitions",
        "parameters": [
          {
            "in": "path",
            "name": "universe",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "language",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JewelleryDefinition"
                  }
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Gets the basic armor definitions most universes need",
        "tags": [
          "universe-creation-service"
        ]
      }
    },
    "/api/{universe}/universe-creation/materials": {
      "post": {
        "operationId": "createDefaultMaterials",
        "parameters": [
          {
            "in": "path",
            "name": "universe",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "language",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "Creates a few basic materials most universes need",
        "tags": [
          "universe-creation-service"
        ]
      }
    },
    "/api/{universe}/universe-creation/secondary-attribute-info": {
      "post": {
        "operationId": "getSecondaryAttributeInfo",
        "parameters": [
          {
            "in": "path",
            "name": "universe",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SecondaryAttributeDTO"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SecondaryAttributeInfo"
                  }
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Creates a few basic materials most universes need",
        "tags": [
          "universe-creation-service"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "JewelleryDefinition": {
        "required": [
          "name",
          "tag"
        ],
        "type": "object",
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "tag": {
            "type": "string"
          }
        }
      },
...
      "SecondaryAttributeDTO": {
        "required": [
          "calculationFormula",
          "consumable",
          "name",
          "shortName"
        ],
        "type": "object",
        "properties": {
          "calculationFormula": {
            "type": "string"
          },
          "consumable": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "shortName": {
            "type": "string"
          }
        }
      },
      "SecondaryAttributeInfo": {
        "type": "object",
        "properties": {
          "average": {
            "type": "integer",
            "format": "int32"
          },
          "max": {
            "type": "integer",
            "format": "int32"
          },
          "min": {
            "type": "integer",
            "format": "int32"
          },
          "notPrecise": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}